Clean up of canon decoder
Former-commit-id: 2ebc82d12b009296079d10a7c3fd156eb170926d [formerly 50a454095212a65459e58ff27600e18663d12aea]
Former-commit-id: 479976c27c804edd77e646e22b47d39907b170fd [formerly c2bcbdb523c44b9359b477362e804d73c27c16b0]
Former-commit-id: b105dda1ac13e88873eae652447055d06690fe5e
Former-commit-id: 7b19c25600860e1b841983c853721a8298339bab

+++ Source/RawParser/Model/Parser/Decoder/Decompressor/LJpegPlain.cs
@@ -243,7 +243,7 @@ namespace RawNet.Decoder.Decompressor
             }
             // Fix for Canon 6D raw, which has flipped width & height
             // see FIX_CANON_FLIPPED_WIDTH_AND_HEIGHT
-            uint sliceH = frame.numComponents == 3 ? Math.Min(frame.width, frame.height) : frame.height;
+            long sliceH = frame.numComponents == 3 ? Math.Min(frame.width, frame.height) : frame.height;
 
             // To understand the CR2 slice handling and sampling factor behavior, see
             // https://github.com/lclevy/libcraw2/blob/master/docs/cr2_lossless.pdf?raw=true
